Implement call
