+++ src/backend/libs/efiling-commons/src/main/java/ca/bc/gov/open/jag/efilingcommons/submission/EfilingReviewService.java
@@ -3,6 +3,7 @@ package ca.bc.gov.open.jag.efilingcommons.submission;
 import ca.bc.gov.open.jag.efilingcommons.submission.models.FilingPackageRequest;
 import ca.bc.gov.open.jag.efilingcommons.submission.models.review.ReviewFilingPackage;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,5 +12,7 @@ public interface EfilingReviewService {
     Optional<ReviewFilingPackage> findStatusByPackage(FilingPackageRequest filingPackageRequest);
 
     List<ReviewFilingPackage> findStatusByClient(FilingPackageRequest filingPackageRequest);
+
+    Optional<byte[]> getSubmissionSheet(BigDecimal packageNumber);
 
 }
