+++ src/backend/libs/efiling-cso-client/src/main/java/ca/bc/gov/open/jag/efilingcsoclient/CsoReviewServiceImpl.java
@@ -4,6 +4,8 @@ package ca.bc.gov.open.jag.efilingcsoclient;
 import ca.bc.gov.ag.csows.filing.status.FilingStatus;
 import ca.bc.gov.ag.csows.filing.status.FilingStatusFacadeBean;
 import ca.bc.gov.ag.csows.filing.status.NestedEjbException_Exception;
+import ca.bc.gov.ag.csows.reports.Report;
+import ca.bc.gov.ag.csows.reports.ReportService;
 import ca.bc.gov.open.jag.efilingcommons.exceptions.EfilingStatusServiceException;
 import ca.bc.gov.open.jag.efilingcommons.submission.EfilingReviewService;
 import ca.bc.gov.open.jag.efilingcommons.submission.models.FilingPackageRequest;
@@ -14,6 +16,7 @@ import org.slf4j.LoggerFactory;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -24,13 +27,16 @@ public class CsoReviewServiceImpl implements EfilingReviewService {
 
     private final FilingStatusFacadeBean filingStatusFacadeBean;
 
+    private final ReportService reportService;
+
     private final FilePackageMapper filePackageMapper;
 
-    public CsoReviewServiceImpl(FilingStatusFacadeBean filingStatusFacadeBean, FilePackageMapper filePackageMapper) {
+    public CsoReviewServiceImpl(FilingStatusFacadeBean filingStatusFacadeBean, ReportService reportService, FilePackageMapper filePackageMapper) {
 
         this.filingStatusFacadeBean = filingStatusFacadeBean;
-
+        this.reportService = reportService;
         this.filePackageMapper = filePackageMapper;
+
     }
 
     @Override
@@ -73,5 +79,18 @@ public class CsoReviewServiceImpl implements EfilingReviewService {
             throw new EfilingStatusServiceException("Exception while finding status list", e.getCause());
 
         }
+    }
+
+    @Override
+    public Optional<byte[]> getSubmissionSheet(BigDecimal packageNumber) {
+        Report report = new Report();
+        report.setName(Keys.REPORT_NAME);
+        report.getParameters().addAll(Arrays.asList(Keys.REPORT_PARAMETER, packageNumber.toPlainString()));
+
+        byte[] result = reportService.runReport(report);
+
+        if (result == null || result.length == 0) return Optional.empty();
+
+        return Optional.of(result);
     }
 }
