+++ src/backend/libs/efiling-cso-client/src/test/java/ca/bc/gov/open/jag/efilingcsoclient/csoReviewServiceImpl/FindPackagesByClientId.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open.jag.efilingcsoclient.csoStatusServiceImpl;
+package ca.bc.gov.open.jag.efilingcsoclient.csoReviewServiceImpl;
 
 import ca.bc.gov.ag.csows.filing.status.FilePackage;
 import ca.bc.gov.ag.csows.filing.status.FilingStatus;
@@ -19,13 +19,13 @@ import org.mockito.MockitoAnnotations;
 
 import javax.xml.datatype.DatatypeConfigurationException;
 import java.math.BigDecimal;
-import java.util.Optional;
+import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@DisplayName("Find Package Test Suite")
-public class FindPackageByIdTest {
+@DisplayName("Review Service Test Suite")
+public class FindPackagesByClientId {
     public static final String CLIENT_FILE_NO = "CLIENTFILENO";
     public static final String COURT_CLASS_CD = "CLASSCD";
     public static final String COURT_FILE_NO = "FILENO";
@@ -41,13 +41,10 @@ public class FindPackageByIdTest {
     FilingStatusFacadeBean filingStatusFacadeBean;
 
     private final BigDecimal SUCCESS_CLIENT = BigDecimal.ONE;
-    private final BigDecimal SUCCESS_PACKAGE = BigDecimal.ONE;
 
     private final BigDecimal EXCEPTION_CLIENT = BigDecimal.TEN;
-    private final BigDecimal EXCEPTION_PACKAGE = BigDecimal.TEN;
 
     private final BigDecimal NOTFOUND_CLIENT = BigDecimal.ZERO;
-    private final BigDecimal NOTFOUND_PACKAGE = BigDecimal.ZERO;
 
     private static CsoReviewServiceImpl sut;
 
@@ -59,41 +56,42 @@ public class FindPackageByIdTest {
         FilingStatus filingStatus =  createFilingStatus();
         filingStatus.getFilePackages().add(createFilePackage());
 
-        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(SUCCESS_PACKAGE), ArgumentMatchers.eq(SUCCESS_CLIENT), any(), any(), any(), any(), any(), any())).thenReturn(filingStatus);
+        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(SUCCESS_CLIENT), any(), any(), any(), any(), any(), any())).thenReturn(filingStatus);
 
-        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(NOTFOUND_PACKAGE), ArgumentMatchers.eq(NOTFOUND_CLIENT), any(), any(), any(), any(), any(), any())).thenReturn(createFilingStatus());
+        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(NOTFOUND_CLIENT), any(), any(), any(), any(), any(), any())).thenReturn(createFilingStatus());
 
-        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(EXCEPTION_PACKAGE), ArgumentMatchers.eq(EXCEPTION_CLIENT), any(), any(), any(), any(), any(), any())).thenThrow(new NestedEjbException_Exception());
+        Mockito.when(filingStatusFacadeBean.findStatusBySearchCriteria(any(), any(), any(), any(), any(), any(), any(), ArgumentMatchers.eq(EXCEPTION_CLIENT), any(), any(), any(), any(), any(), any())).thenThrow(new NestedEjbException_Exception());
 
-        sut = new CsoReviewServiceImpl(filingStatusFacadeBean, new FilePackageMapperImpl());
+        sut = new CsoReviewServiceImpl(filingStatusFacadeBean, null, new FilePackageMapperImpl());
     }
 
-    @DisplayName("OK: package found")
+    @DisplayName("OK: packages found")
     @Test
     public void testWithFoundResult() throws DatatypeConfigurationException {
-        Optional<ReviewFilingPackage> result = sut.findStatusByPackage(new FilingPackageRequest(SUCCESS_CLIENT, SUCCESS_PACKAGE));
+        List<ReviewFilingPackage> result = sut.findStatusByClient(new FilingPackageRequest(SUCCESS_CLIENT, null));
 
-        Assertions.assertEquals(COURT_FILE_NO, result.get().getCourt().getFileNumber());
-        Assertions.assertEquals(COURT_CLASS_CD, result.get().getCourt().getCourtClass());
-        Assertions.assertEquals(COURT_LEVEL_CD, result.get().getCourt().getLevel());
-        Assertions.assertEquals(COURT_LOCATION_CD, result.get().getCourt().getLocationCd());
-        Assertions.assertEquals(COURT_LOCATION_NAME, result.get().getCourt().getLocationName());
+        Assertions.assertEquals(1, result.size());
+        Assertions.assertEquals(COURT_FILE_NO, result.get(0).getCourt().getFileNumber());
+        Assertions.assertEquals(COURT_CLASS_CD, result.get(0).getCourt().getCourtClass());
+        Assertions.assertEquals(COURT_LEVEL_CD, result.get(0).getCourt().getLevel());
+        Assertions.assertEquals(COURT_LOCATION_CD, result.get(0).getCourt().getLocationCd());
+        Assertions.assertEquals(COURT_LOCATION_NAME, result.get(0).getCourt().getLocationName());
 
     }
 
     @DisplayName("Ok: no packages found")
     @Test
     public void testWithNoResult() {
-        Optional<ReviewFilingPackage> result = sut.findStatusByPackage(new FilingPackageRequest(NOTFOUND_CLIENT, NOTFOUND_PACKAGE));
+        List<ReviewFilingPackage> result = sut.findStatusByClient(new FilingPackageRequest(NOTFOUND_CLIENT, null));
 
-        Assertions.assertFalse(result.isPresent());
+        Assertions.assertTrue(result.isEmpty());
 
     }
 
     @DisplayName("Exception: filing status facade throws an exception")
     @Test
     public void testWithException() {
-        Assertions.assertThrows(EfilingStatusServiceException.class, () -> sut.findStatusByPackage(new FilingPackageRequest(EXCEPTION_CLIENT, EXCEPTION_PACKAGE)));
+        Assertions.assertThrows(EfilingStatusServiceException.class, () -> sut.findStatusByClient(new FilingPackageRequest(EXCEPTION_CLIENT, null)));
     }
 
     private FilingStatus createFilingStatus() {
